Network.Useless

Description

Useless is a Module for building Webapplications

Documentation

data HTTPRequest

The HTTP Request

Constructors

HTTPRequest

Fields

httpReqMethod :: String

httpReqVersion :: String

httpReqFile :: String

httpReqQueries :: [(String, String)]

httpReqHeader :: [(String, String)]

data HTTPResponse

The HTTP Response

Constructors

HTTPResponse

Fields

httpResStatus :: Integer

httpResHeader :: [(String, String)]

httpResBody :: String

initUseless :: IO Useless

initUseless initializes a Useless server layout. It has to be called befor all register calles

register :: Useless -> String -> UselessSite -> IO ()

register adds resources to your a server layout, and returns a new server layout.

The given UselessSite function is called, whenever someone requests the resource.

e.g:

  server = do 
  	layout <- initUseless
  	register layout "/test" someFunction

createBasicHTTP :: String -> HTTPResponse

createBasicHTTP creates a very Basic HTTP response from a given String

createErrorResponse :: Integer -> UselessSite

createErrorResponse returns a UselessSite function , for a given HTTP Status code

e.g:

  	register myServer "/forbidden" (createErrorResponse 403)

addToUseless :: Useless -> String -> String -> IO ()

addToUseless adds a Key Value Pair to a shared Useless object

If the key already has a value, it will be overwritten.

e.g:

import Network.Useless
import Maybe

main = do
       theServer <- initUseless
       register theServer "/" showMe
       register theServer "/change" testadd
       startServer 8080 theServer

showMe :: UselessSite
showMe useless request = do
       val <- requestFromUseless useless "ThisSite"
       return $ createBasicHTTP $ fromMaybe "Not Visited" val

testadd :: UselessSite
testadd useless request = do
        addToUseless useless "ThisSite" "Visited"
        return $ createBasicHTTP $ "Nichts zu sehen"

startServer :: Integer -> Useless -> IO ()

startServer starts the server at a specific Port

type UselessSite = Useless -> HTTPRequest -> IO HTTPResponse

UselessSite is the type for your Useless functions.

   testsite :: UselessSite
   testsite useless request = return $ HTTPResponse {httpResStatus = 200, httpResHeader=[("foo:", "Bar,Baz")], httpResBody="Test123"}

data UselessData

UselessData consists of the sites, that can be served by your Server and a String Map, that can be used for Shared memory (Future versions will expand this)

Constructors

UselessData

Fields

sites :: Map String UselessSite

stringmap :: Map String String

getUselessData :: Useless -> IO UselessData

getUselessData returns the UselessData associated with a Useless shared memory

requestFromUseless :: Useless -> String -> IO (Maybe String)

requestFromUseless looksup a String in the shared Memory. If no such Value for the Key is found, Nothing is returned.

see addToUseless for an example.