Useless

Description

Useless is a Module for building Webapplications

Documentation

data HTTPRequest

The HTTP Request

Constructors

HTTPRequest

Fields

httpReqMethod :: String

httpReqVersion :: String

httpReqFile :: String

httpReqQueries :: [(String, String)]

httpReqHeader :: [(String, String)]

data HTTPResponse

The HTTP Response

Constructors

HTTPResponse

Fields

httpResStatus :: Integer

httpResHeader :: [(String, String)]

httpResBody :: String

initUseless :: IO (Map String (HTTPRequest -> IO HTTPResponse))

initUseless initializes a Useless server layout.

register :: String -> (HTTPRequest -> IO HTTPResponse) -> Map String (HTTPRequest -> IO HTTPResponse) -> IO (Map String (HTTPRequest -> IO HTTPResponse))

register adds resources to your a server layout, and returns a new server layout.

The given function is called, when the resource is requested. It gets a HTTPRequest and has to respond with a HTTPResponse

e.g:

  server = do 
  	layout <- initUseless
  	layout <- register "/test" someFunction layout 

createErrorResponse :: Integer -> HTTPRequest -> IO HTTPResponse

createErrorResponse returns a function, that gets a HTTPRequest and answers with a HTTPResponse, for a given HTTP Status code

e.g:

  	layout <- register "/forbidden" (createErrorResponse 403) layout 

startServer :: Integer -> Map String (HTTPRequest -> IO HTTPResponse) -> IO ()

startServer starts the server at a specific Port